requirements.txt
crewai>=0.86.0
crewai-tools>=0.17.0
python-pptx>=1.0.2
streamlit>=1.38.0
python-dotenv>=1.0.1